refactor (again) and finally make pylint shut up
